Renamed event-emitter to eventable.
